+++ NEWS.rst
@@ -6,6 +6,8 @@ every change, see the Git log.
 
 Latest
 ------
+* Minor: Use gradle 4.4.
+* Major: Move arm and armv7 libraries into armeabi-v7a jni folder.
 * Major: Moved ``MAX_SYMBOL_SIZE`` and ``MAX_GENERATION_SIZE`` from ``Source``
   into ``AutoSource`` and ``ManualSource``.